remove split package
