+++ apollo-normalized-cache-api/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/api/NormalizedCache.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.cache.normalized
+package com.apollographql.apollo3.cache.normalized.api
 
 import kotlin.jvm.JvmStatic
 import kotlin.jvm.JvmSuppressWildcards
