+++ apollo-normalized-cache-api/src/commonTest/kotlin/com/apollographql/apollo3/cache/normalized/CacheKeyResolverTest.kt
@@ -6,6 +6,8 @@ import com.apollographql.apollo3.api.Executable
 import com.apollographql.apollo3.api.ObjectType
 import com.apollographql.apollo3.cache.normalized.CacheKeyResolverTest.Fixtures.TEST_LIST_FIELD
 import com.apollographql.apollo3.cache.normalized.CacheKeyResolverTest.Fixtures.TEST_SIMPLE_FIELD
+import com.apollographql.apollo3.cache.normalized.api.CacheKey
+import com.apollographql.apollo3.cache.normalized.api.CacheKeyResolver
 import com.apollographql.apollo3.exception.CacheMissException
 import kotlin.test.BeforeTest
 import kotlin.test.Test
