--- apollo-normalized-cache/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/CacheContext.kt
@@ -1,3 +1,0 @@
-package com.apollographql.apollo3.cache.normalized.internal
-
-
